Update patch set 1
Patch Set 1:

Adam, does this mean you or the Go team considers stream ciphers to be obsolete?  If so I would strongly contest that in general; they are a very general and widely useful cryptographic building block when used with care in larger protocols - e.g., see the ways they' re used in anonymity systems such as Tor or Dissent.  For plain, run-of-the-mill encryption/decryption for two-party communication I agree that AEADs are the safer and stronger primitive to move to, but there are a lot of encryption purposes AEADs can't serve - such as protocols (like Tor or Dissent) that must apply multiple layers of encryption to the same plaintext without expanding it at each level.  You simply can't use AEADs in these kinds of scenarios for various reasons such as because the authenticators are only secure in two-party secret-sharing scenarios, the authenticators would leak identity information for traffic analysis, etc.  I'm just saying that it seems premature to label cipher.Stream or stream ciphers in general as "deprecated".

Patch-set: 1