ssh: The SSH server does not allow for setting a version string in the same manner as the client.
This update adds a ServerVersion member to the ServerConfig structure
which when set, causes the server to use that member instead of the
default version string.  This allows building an golang based SSH
server which can present any version string the user wishes.

It also adds an else statement to the client assignment of the
ClientVersion to avoid an allocation when using a user defined
ClientVersion.

Change-Id: I43d97cfd5a174f2c68f53c5b4e267539ef21937b