Update patch set 1
Patch Set 1:

The use case I am shooting for is presenting a version string that certain non-standard clients are hard wired to look for.  Basically allowing a golang based SSH server to look like something else to a casual observer.

I am also using the library as part of a honeypot and need to present a banner that looks like openssh, dropbear, etc...

I am not sure what the implications of parking the version string in the central config and using it for both would be on other packages, so I will defer to others.  If others decide that is a better route I will re-roll the patch to use that and resubmit.

Patch-set: 1