Update patch set 3
Patch Set 3:

> then I think this review site has problem to me

If you load https://go-review.googlesource.com/#/c/1861 then you should see reply in the first gray bar across the screen. (It's not linked to a specific message in the UI.)

 > case len(t.line) == 0 && t.cursorX == 0 && t.cursorY == 0:
 > return nil

Ok. Could you also change the "fallthrough" just above to be "return 0"? The latter is also one line but there's a tiny reduction in the chance that someone could miss the "fallthough" in future and have a bug.

Patch-set: 3
Label: Code-Review=0