Update patch set 1
Patch Set 1:

I have poked around a few forums and spoke to some people that use the crypto/ssh package and have gotten a little push back on moving the version string to the general config as it would force code change and break compatibility with the current method of specifying version strings for clients.  As far as maintaining compatibilty it seems best to add the additional member to the server config and not move the client member.  I will leave the patch as is unless told otherwise.

Patch-set: 1
Label: Code-Review=0