Update patch set 1
Patch Set 1:

I don't see a clear reason to put cityhash in the standard library.  There are go gettable packages that provide it: http://godoc.org/?q=cityhash .

The reason to put it in the standard library would be if it is needed by something else in the library, such as the net/http package.

Patch-set: 1