Update patch set 3
Patch Set 3:

I'm not clear, from the bug description, what the underlying problem is here. Calling advanceCursor(1) n times should be equal to advanceCursor(n). If not, then there's a bug in advanceCursor and this is just papering over the problem.

The advanceCursor function does very a special case for handling when the cursor ends up at the end of a line, but it's not clearly broken for the multi-line case. (Although terminals are so complex and badly specified that I can believe that it is.)

(Note: I'm unable to play with this myself until I get back from holiday.)

Patch-set: 3