Update patch set 3
Patch Set 3:

> I'm not clear, from the bug description, what the underlying
 > problem is here. Calling advanceCursor(1) n times should be equal
 > to advanceCursor(n). If not, then there's a bug in advanceCursor
 > and this is just papering over the problem.
 > 
 > The advanceCursor function does very a special case for handling
 > when the cursor ends up at the end of a line, but it's not clearly
 > broken for the multi-line case. (Although terminals are so complex
 > and badly specified that I can believe that it is.)
 > 
 > (Note: I'm unable to play with this myself until I get back from
 > holiday.)

I don't think that can be fixed in advanceCursor unless we change it to the logic like in writeLine, to handle one termWidth at most and insert a line end '\n' once it reached termWidth every time; but in this case writeLine is the higher level function served in this purpose, we don't need another function to behave like writeLine;

func (t *Terminal) writeLine(line []rune) {
	for len(line) != 0 {
		remainingOnLine := t.termWidth - t.cursorX
		todo := len(line)
		if todo > remainingOnLine {
			todo = remainingOnLine
		}
		t.queue(line[:todo])
		t.advanceCursor(visualLength(line[:todo]))
		line = line[todo:]
	}
}

in the case of eraseNPreviousChars, if it has already queued number n of spaces longer than termWidth, it's already broken that we don't have a chance to insert '\n' before the spaces reached termWidth. So it is eraseNPreviousChars need to be fixed, not advanceCursor.

Patch-set: 3