Update patch set 8
Patch Set 8:

I usually take the view that more testing is better than less; none of the tests are particularly expensive.

But the point re log-spam well taken. I'll prune that, substantially. (In particular, I see no reason to note the *success* of any particular instance; it is only interesting if an instance fails.)

Patch-set: 8