openpgp: Only call Decrypt on a copy of SymmetricKeyEncrypted structure.
Since the Decrypt() method of the SymmetricKeyEncrypted type changes the
state of the object in various ways, including attempting the decryption
the encrypted key in place, the said object cannot subsequently be used
in other decryption attempts with other passphrases. This means that if
the first attempt to read a message fails due to a wrong passphrase,
subsequent attempts will never succeed, even when the correct passphrase
is used, since a) The SymmetricKeyEncrypted object will have been marked
as already decrypted as a result of the first Decrypt() call. b) The
original encrypted passphrase will have been overwritten by the first
Decrypt() call.

Fix this by only calling Decrypt() on a pristine copy of the original
SymmetricKeyEncrypted object.

Fixes #9315

Change-Id: Id2c5351150641a2ee9133e0b74d039d8e157d0e7