ssh: expose the local and remote ids of the channel.
SSH exposes local and remote ids for channels that are useful in keeping
track of channels. This is useful if we want to implement
something like opening a channel of a channel on the client side (e.g.
opening an ssh channel, then opening another ssh channel based on the
existing channel).

For an example usecase, see github.com/progrium/duplex, where we use these
ids to connect channels with each other.

Change-Id: I6388220cf27839e253753031b380f2e63f66c40a