Update patch set 3
Patch Set 3:

I worry here that we're adding complexity to work around an original mistake.

Since the SymmetricKeyEncrypted can't figure out whether the decryption was good or not it was probably a mistake in the first place for Decrypt to mutate the object. Decrypt should probably have returned the key instead.

Patch-set: 3