Update patch set 2
Patch Set 2:

Yes, it's fair to say the interface is already too big and that it should maybe be a struct. But if it's going to be redesigned, why not let this patch through? It would help ensure this feature is included when it's redesigned.

And saying we should use our own ID for the exact same purpose of the ID the spec tells us is available is forcing us to design a redundant protocol around this library as opposed to the SSH spec. 

This is such a minor addition that Makes Sense. I'd love to see it get through.

Patch-set: 2