Update patch set 2
Patch Set 2:

These seem like reasonable points but are ultimately arbitrary or subjective, so I feel compelled to counter for sake of balanced perspective. 
 > 
 > * You are implementing an application on top of the SSH protocol.
 > Use the application level mechanisms; channel IDs aren't.
The SSH Connection Layer RFC makes no distinction of "application layer" when talking about custom types, so such a distinction is arbitrarily made by a library author or maintainer. This library is better than most at exposing the extensible mechanisms of the Connection Layer, but to stop at channel ID is an arbitrary, subjective decision of the library design. I can respect this, but I just want to point out that a library that does expose channel ID is still fully compliant with the protocol spec.
 > 
 > * There is no point in exposing the IDs if you have only one
 > channel. My suspicion is that you want to send the ID of an SSH
 > channel over a separate communication mechanism (maybe another SSH
 > channel?). This is unsound.
Yes, it is used in referencing channels from other channels. Exposed channel IDs are being used in an extension of the library/protocol, so other mechanisms are used to prevent the situation you described. Both channels can communicate a reference to each other, and if the other does not exist with that reference, the relationship or entire channel can be discarded. As pointed out before, this reference ID could be something other than channel ID, but it also *can* be channel ID and the only reason it can't be is if the library implementing Connection Layer prevents it. 
 > 
 > yes, but it's also an addition that adds little usefulness, and
 > makes the package larger and therefore harder to understand for
 > users.
Technically, yes it makes it larger. But anecdotally this is already one of the most succinct and understandable libraries for SSH, and one of the few that properly exposes the extensible nature of the Connection Layer. I understand in principle how you can say exposing channel IDs makes it harder to understand. But from a practical perspective, I assure you, it won't. However, I appreciate the principled approach to design in the Go community and realize design is not always a matter of objectivity.

Patch-set: 2
Reviewer: Gerrit User 6035 <6035@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0