crypto/bcrypt: Disallow null characters in passwords
"abc\x00def" and "abc\x00ghi" are silently treated as the same password, when
clearly they are not. This is especially problematic when bcrypt is fed the
raw output of another hash function.

This change will only trigger an error when generating new digests from
passwords containing null bytes. Should comparisons, and thus potentially
existing sign-ons fail as well?

Change-Id: I1bc0f386964969c121f42f2f32b0d367fa25461c