crypto/ssh: fix ssh-conform encoding of certs with critical options
Attention - BREAKING change for the certificates generated with
the previous versions of crypto/ssh!  Need to regenerate
certificates with a version of crypto/ssh library including
this fix.

[PROTOCOL.cerkeys] requires two length fields for non-empty
values of critical options (or extensions - but those are
currently always empty)  - see
https://bugzilla.mindrot.org/show_bug.cgi?id=2389.
Added special handling of those values as marshalTuples and
parseTuples as well as related tests.

Fixes #10569

Change-Id: Iecbfca67a66668880635141c72bc5fc370a9c112