Update patch set 1
Patch Set 1:

I've done a significant pass on this. Please check to see if I've obviously messed anything up.

One blocker is that rc2/rc2.go lists an MIT license. In order to ensure the licensing is sound I'll email dgryski and ask him to chime in here to confirm that the code is submitted under the CLA.

Patch-set: 1