x/crypto/openpgp: found infinite loop problem on openpgp symmetric decryption.
When decrypting some content with a bad passphrase, the prompt function
should return an error after some time (timeout, user decision of quitting)
else openpgp.read enters an infinite loop. Code is left as is but a test
illustrates what to do to avoid the problem.

Some simple sample code that shows the bug is available here :
https://gist.github.com/ufoot/b4beabbc9050df4217c2

Change-Id: I245241f5bc4f85f0b70b78f6441c6e81705ab7c8