Update patch set 3
Patch Set 3:

I've been playing a bit with building the assembly code with Go 1.4.2, but the assembler pretty much chokes on it.

One issue is that pre Go 1.5, the assembled cannot handle line comments within multi-line defines. This can be easily solved by replacing them with block comments.

A second issue is that it cannot handle $8*25 evaluations, but this can also be fixed with $(8*25).

A third issue is that it needs the semicolons after the end of each line in a multi-line macro, which is easily fixable.

However, with all the above fixes included, the macro definition seems to be too large for the assembler, as three quarters through the macro, the compiler simply errors out with `syntax error, last name: <a few random chars>`, where the characters are some unprintable chars, always different at every run.

Given the seeming bug in the compiler, I'm not sure if we should try and fix up the assembly code to make it backward compatible, or rather to add a build constraint requiring Go 1.5? Any thoughts on this?

Patch-set: 3