Update patch set 3
Patch Set 3:

@Peter: IMO Go 1.4 compatibility is a must. I would expect it to be used in productions.

@minux, bradfitz:

Admittedly 20% isn't that impressive, but I think for some applications (storage, etc) this would be a welcome improvement, since hashing can often be CPU limited. And a throughput of 170 instead of 140MB/s would be a welcome upgrade.

Since ssa is about half a year from production release AFAIK, I think it would be ok to include, but with a note to check performance improvement after Go 1.7 releases.

I *may* be able to get the speedup closer to 30%, with some ideas I have. Would that be worth including?

Patch-set: 3