Update patch set 4
Patch Set 4:

@Klaus PTAL

I've fixed all the quirks of the Go 1.4 assembler. The segfault was caused by the round macro being too large in length and probably overflowing some internal buffer of the assembler. I've reduced the length of the comments within the macro, which cut a ton of whitespace off the formatted code, fitting it into said buffer. Now the code compiler and passes all tests on Go 1.4.2, 1.5.2 and also Go tip.

Patch-set: 4