ssh: parse cert-authority known_hosts
@cert-authority known_hosts format is slightly different than the
authorized_keys format.

from the sshd manpage:
  @cert-authority *.mydomain.org,*.mydomain.com ssh-rsa AAAAB5W

where "*.mydomain.org,*.mydomain.com" are the domains for which the
given certificate can sign host keys.

This is probably the wrong way to do this (please suggest alternatives),
but this change returns "*.mydomain.com,*.mydomain.org" as the options
return value and then parses the rest of the certificate.

Those domains should be taken into account when doing certificate, but I
didn't see a way that made sense to make them available to
CertChecker.CheckCert(). That's part of why I don't think this is the
right way to it.

Change-Id: I9258ecf2b38258e31bcb6e73ac042ad8125fd2d1