ssh: KnownHostPrincipal aids checking if a host is in the known_hosts file.
When using host certificates, the list of hosts/domains for which a given
certificate are valid are stored in the known_hosts file along with the
publickey. This function consolidates most of the logic for matching a presented
principal against a know_hosts entry.

I haven't found the right way to plug this directly into the ClientConfig or
CertChecker types, suggestions are welcome.

Change-Id: I415b45b677d11dcc92ea96ee7df5ec2f5560bb3c