Update patch set 3
Patch Set 3:

(1 comment)

> I fixed a very, very minor things.
 > 
 > However, the encoding in pkcs12.go isn't really viable. The shallow
 > nature of safeBag and other types may have made sense for parsing
 > but for serialisation new types might be needed.
 > 
 > Also, while parsing PKCS#12 files is a plausible use since they do
 > exist, they should be strongly discouraged. Supporting their
 > serialisation seems counter to this.

Thank you for the review, and fixes - I clearly missed the padding sweep.

Encoding was written with the intent to interchange .pfx files, which is narrow. The parsing code is heavily biased towards this fact, and likewise so is this serialization code.  Decode and encode are now a matched, but limited pair.

Please let me know if you think this is acceptable or not.  I would prefer to share, but I can retract the PR, and go with an repo internal approach.

Patch-set: 3
Reviewer: Gerrit User 12226 <12226@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0