Update patch set 2
Patch Set 2:

> (2 comments)

Changes sound good; I wasn't sure about a test because of how go test handles terminals.  I did confirm on windows and darwin I could get this test to run and pass (not just skipped) if using the `-v` option (and that it would previously fail for windows).  If you don't pass -v I think os.Stdout is redirected and so this test would just be skipped; it is better than nothing though.

Patch-set: 2