x/crypto/openpgp/encoding: add Field, MPI & BitString types
The Field interface represents an encoded packet field. MPI is a
Multiprecision Integer field as defined in rfc 4880, Section 3.2.
BitString is a variable length field as defined in rfc 6637, Section 8.

By using the MPI and BitString types for packet fields, the
packet.ecdsaKey and packet.ecdhKdf types are replaced with the
parseECDSA and parseECDH methods of packet.PublicKey.

Change-Id: I8a7c16f8ab77c0aad3ed37540e606d9285907a5d