x/crypto/openpgp/internal/encoding: add Field, MPI & OID types
The Field interface represents an encoded packet field. MPI is a
Multiprecision Integer field as defined in rfc 4880, Section 3.2.
OID is a variable length field as defined in rfc 6637, Section 8.

By using the MPI and OID types for packet fields, the packet.ecdsaKey
and packet.ecdhKdf types are replaced with the parseECDSA and parseECDH
methods of packet.PublicKey.

Change-Id: Id4c79a85c095f0fab02fbf084a81deadd1dcbe32