Update patch set 6
Patch Set 6:

> (Sorry if this is the wrong place for such a question)

After pondering it for a little while, I've added PublicKey and PrivateKey types to this package, although they're just []byte underneath and so those who want the more modern, byte-orientated interface should not be too inconvenienced.

I'll leave it a couple of days to solicit opinions about this before landing.

Patch-set: 6