Update patch set 5
Patch Set 5:

It seems that this broke the integration tests under ssh/test/ , leading to 

		debug1: SSH2_MSG_NEWKEYS received
		Bad packet length 1560169096.

ie. it will probably break anyone who runs long-lived connections with openssh.
 
Is there a way that we can make the builders run the integration tests too on platforms that support it? Just Linux x86/amd64 would be helpful.

Patch-set: 5