Update patch set 2
Patch Set 2:

The only thing to know about Gerrit is that the Change-Id value is the primary key and the workflow you need to use for revising a change in review is to keep squashing your changes back into a single commit (generally with git commit --amend), always making sure you don't mess up your original Change-Id line.

Outside of that, there's nothing Gerrit-specific to learn and it's just learning git.

Patch-set: 2