Update patch set 6
Patch Set 6:

It appears that at least one of the integration tests fails probably due to a change in OpenSSH.  I tried in a clean install of OpenSSH_7.2p2 on fedora 23 (using docker) and saw the same problem.  Trying again with fedora 22 (again in docker) which has OpenSSH_6.9p1 enabled the tests to pass against master. I don't understand the difference.  I'm now using my F22 docker image as a basis to try and work on ensuring the tests are modified where needed to cover the ed25519 code.

Patch-set: 6