Update patch set 6
Patch Set 6:

over here I see a problem with TestCertLogin, which might be because of some google specific extensions.

I'd recommend adding a TestKeyAlgorithms to session_test.go, similar to the MAC, Kex and cipher tests, which does a client login using each of the key types. You can make sure it passes as expected by removing the fallbacks.

If there are other unrelated failures under test/, you can ignore them. I just want to be sure that the code you want to submit works with the reference implementation.

Patch-set: 6