Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Thanks for the suggestion, it makes sense to me.
> 
> I would like to go slightly further however.  I would prefer two things to be tested.  First the use of various types of key by the golang client to authenticate against the openssh server as you suggested adding, but also the use of the same types of keys by the openssh server as host keys, verified by the golang client.  I don't think the current code does this now for all the current types.  I'd like to split out the current PEMData map (which is mostly host keys, with one "user" key) into two maps - one for host key testing and one for client key testing, and update the current tests to more fully test the current code on master.   I think this is a big enough change to do first in a separate change before re-basing this work onto that.
> Is that an acceptable approach?

I don't think the hostkeys need to be separately tested. The code for doing hostkeys is generic, so it should already be tested with the existing code. 

I want to have some OpenSSH coverage is mainly there to prevent screwups like 

 https://go-review.googlesource.com/c/22418/

where the unittests were AOK, but it actually didn't work in practice because the unittest tests the golang code against itself

Patch-set: 6
Label: Code-Review=0