Update patch set 1
Patch Set 1:

(1 comment)

> (1 comment)
 > 
 > I don't think adds much value.

I just think that it makes sense to implement this message so that any consumers of the ssh library wouldn't have to if they wanted to send these. As a consumer I may know my height/width but wouldn't know without looking up the RFC document that a) this message exists and can be sent, b) that you send the height/width based on columns/rows and by pixels.  Just not sure why we wouldn't put this in here with the other message type implementations.

Patch-set: 1
Label: Code-Review=0