Update patch set 1
Patch Set 1:

This looks like a mailing list message in a commit message with no files changed?

Are you trying to send a change, or a message?

I'm going to close this for now and you can retry if your intention was to use Gerrit to send a review.

Patch-set: 1
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0