Update patch set 1
Abandoned

This looks like a mailing list message in a commit message with no files changed?

Are you trying to send a change, or a message?

I'm going to close this for now and you can retry if your intention was to use Gerrit to send a review.

Patch-set: 1
Status: abandoned