Update patch set 5
Patch Set 5: Code-Review+2

I agree that this is not easy to test. Perhaps with a number of synchronisation points in the code that tests could use to step the state machines in precise order. However, I think it's more important to get the fix in first.

Patch-set: 5
Reviewer: Gerrit User 5425 <5425@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2