opengpg: NewEntity set SelfSignature.PreferredHash if DefaultHash is passed
Say the user want to create a NewEntity, then Serialize the private and public
keys. Then call Encrypt. The Hash used first for Encrypt will be `RIPEMD160`,
which is not compiled in by default. So say they _specifically_ pass
crypto.SHA256 as the Default hash to NewEntity, this should set the
PreferredHash in the SelfSignature so they can be encrypted with that hash.

Related to golang/go#12153

Signed-off-by: Jess Frazelle <me@jessfraz.com>

Change-Id: I3d89f9d60f9abb29ec8365f81da38b8cd0594587