Update patch set 1
Patch Set 1:

> Perhaps I'm missing something, but I don't think I follow.
 > Keyboard-interactive does not seem to have any kind of inherent
 > looping to it. The server sends N questions and expects N answers.

Sure. But the code in client_auth.go has a "for {" loop that means that the server can send many userAuthInfoRequestMsg, each with an array of questions.


 > When used for things like PAM/challenge-response; this means that
 > the server sends *1* prompt, and expects *1* answer. When that
 > answer is wrong, there is no looping that takes place;

Are you sure? The code certainly looks like it'll call the callback multiple times if the server requests another set of answers. Maybe the server is responding with a failure message and another initiateMsg is needed?

 > keyboard-interactive is removed from further consideration, and
 > cannot be retried. If you change the hardcoded username/password/hostname/port
 > parts of the snipped here: https://play.golang.org/p/F-od_urg3G
 > with openssh's sshd configured with
 > UsePAM yes
 > ChallengeResponseAuthentication yes
 > PasswordAuthentication no
 > You can see this behavior. The server sends a single prompt:
 > "Password: ". When the user enters the wrong password, x/crypto/ssh
 > considers the keyboard-interactive method "done" and if no other
 > authentication methods remain, the authentication fails.

If the server is failing the method rather than asking another question then I can see the problem. But it still seems to me that providing auth methods that retry is probably a cleaner way to do this.

Patch-set: 1
Label: TryBot-Result=0