Update patch set 1
Patch Set 1:

> 
 > If the server is failing the method rather than asking another
 > question then I can see the problem. But it still seems to me that
 > providing auth methods that retry is probably a cleaner way to do
 > this.

There's never an opportunity for the server to ask another question because the client is no longer considering the method valid. 
The server sends 1 prompt, the user sees this prompt, and provides their response; that answer is sent to the server. The server responds that it was the wrong answer, this causes auth.auth to return false, []string{<methods>}, nil because the answer was wrong. There wasn't an "error", so we don't bail on line 36, the answer was wrong, so we don't bail on line 40, and so on line 42, tried[auth.method()] = true

Note that there is a flip-side to this; On servers that allow PasswordAuthentication, it is *always* in the list of "methods that may continue" until the server drops you for too many authentication failures; so you need some way of stopping after some "N" prompts. It is also possible (as is the case in the example config above, which is fairly common) for keyboard-interactive auth to do the same thing. So in both cases, you need some way of signaling x/crypto/ssh that we, the client, want to stop trying this method. That's why in the patch I used a sentinel error; so that the consumer of x/crypto/ssh can elect to stop trying an auth method at any point; even when it remains in the list of authentication methods that may continue.

Patch-set: 1