Update patch set 1
Patch Set 1:

I think I glossed over part of your response where you highlighted part of the issue which was the expectation that the server would issue another msgUserAuthInfoRequest. It didn't click in my head because I'm only familiar with how openssh handles this, so it didn't occur to me that it might be different elsewhere. here's the full details of what is happening with openssh: After applying this patch for debugging purposes (because delve hates me) https://bpaste.net/show/b68b39006dfd
I get this when running the example snippet I linked previously:
https://bpaste.net/show/fed7aec53c5a
on the client and https://bpaste.net/show/a3fdc1e03cf4 on a server in debug. (ubuntu 15.04 in a vagrant I happened to have handy)

the server gets the wrong answer, and responds with msgUserAuthFailure not another msgUserAuthInfoRequest. As far as I understand it; an openssh server would only ever send another msgUserAuthInfoRequest if it had more prompts as part of the same challenge-response set (e.g. a "chat" script type thing)

Patch-set: 1