chacha20, poly1305: add ChaCha20 stream cipher and ChaCha20Poly1305 AEAD
Add the ChaCha20 and the ChaCha20Poly1305 AEAD specified in RFC 7539.
Therefore introduced a special poly1305.Hash struct
to avoid memory allocations (for the AEAD).
The chacha20 package is similar to the existing salsa20 package. There is
still room for performance improvements (AVX2...), but maybe it can be used as
a template.

Change-Id: I22713aa955c57a61eeb193d514ad3c7ad832f88d