Update patch set 3
Patch Set 3:

(4 comments)

> (1 comment)
 > 
 > it's a little bit clumsy, because you'll always have to use a
 > closure, but maybe that's ok for an advanced feature like this.

I used a closure in the test-cases just to make it easier to retry with different values; It wouldn't normally be required.

I've added documentation; and added an example test-case (one that doesn't make use of any variables in the environment of the closure) for keyboard-interactive.

If I have misunderstood your statement about the closure let me know. 
Also, If the docs are too little/much I'm happy to revise. Thanks!

Patch-set: 3
Label: Code-Review=0