Update patch set 5
Patch Set 5:

> No test addition or update?

This was a definitional bug and a test would really just have to duplicate the definition. In crypto/tls we have some golden tests and, had we something similar here, they would have shown a difference. But they're really about testing against a different implementation and are frankly quite costly so I'm not enthused about adding them to other packages unless the benefits are clear.

Patch-set: 5