Update patch set 5
Patch Set 5:

rsc.io/letsencrypt's manager has a SetHosts method which allow one to limit the domains that a certificate can be requested for to a list. Perhaps we should include that, maybe with SAN support (make the argument type [][]string, where each []string is a grouped set of domains for the SAN)?

Patch-set: 5
Label: Code-Review=0