Update patch set 6
Patch Set 6:

(2 comments)

> (1 comment)

You said "It would be nice to have another (or use alternative to this) set of X/Y values to verify the padding when you're constructing the EC JWK."

This set of X/Y values is from a RFC and verifies the padding because the X is 65 bytes so a padding of one byte. I couldn't find any other. Do you still want me to revert?

Patch-set: 6
Label: Code-Review=0