Update patch set 2
Patch Set 2:

Sure, 65 bytes makes sense, but:

    echo -n AHKZLLOsCOzz5cY97ewNUajB957y+C+U88c3v13nmGZx6sYl/oJXu9A5RkTKqjqvjyekWF+7ytDyRXYgCF5cj0Kt | base64 -d | wc -c

results in 66, not 65. So this value is most likely already padded. If so, you'll need to remove leading 0s, re-encode to base64 and use that for base64X value.

Another way to find another X which starts from 0s is something like this:

    while true; do openssl ecparam -name secp256k1 -genkey -noout -out ecpriv.pem && openssl ec -in ecpriv.pem -text -noout | grep 04:00 && break; done

When it stops, you should see "04:00:..." and have ecpriv.pem on disk, from which to pick X/Y. But it's probably not worth it for this test.

Patch-set: 2
Label: Code-Review=0