nacl/secretbox: add Seal, Open example
I read the docs and wasn't sure how to use the package - what is the best
way to generate a nonce? Do I need to use the same nonce for encrypting and
decrypting the same message? I read through some code that used this package,
and the tests, found the answers I was looking for, and used them to generate
an example, which hopefully will be helpful to other users of this package.

The one thing I am worried about is the use of a fixed nonce; this is similar
to the approach in the math/rand example.

Change-Id: I24f9682fecb6632f34b4d8d13fed6eb002993677