Update patch set 10
Patch Set 10:

My friend and I wrote a utility like this, using the xenolf/lego package, and here are some suggestions based on our experience.

1) Consider defining the Cache API in terms of raw bytes not certificates. Reason: this lets you write pass-through wrappers like "EncryptedCache", that transparently encrypts the cert data, and forwards it through to an underlying implementation. It seems like the composition of something like this and a DirCache will be valuable, and that'd be complicated (impossible?) to write for an API taking Certificates.

2) Consider adding a Delete method to the Cache API and also using it to store the SNI challenge certs. In our little hobby project, we had frontends stick the acme.invalid certs in etcd, that way if the validation request from LE hit another frontend, it would lookup unknown certs in etcd, and complete the validation that way.

I believe these two changes should handle the multi-manager use-case cleanly, and would let us eventually transition to using this library.

You might want to call this thing a Store instead of a Cache (and adding an actual cache in front of the store that keeps the parsed cert around, so regular cert-serving isn't punished with repeated serialization and deserialization).

Patch-set: 10