Update patch set 11
Patch Set 11:

Forgot about the cache context: wouldn't it make more sense for the implementations of the Cache to have some kind of a cache deadline field instead?

I could imagine a future version of the Manager have a cache timeout but then it wouldn't be fair to cut all kinds of caches under the same deadline. For instance, a hypothetical network cache would have a much longer timeout than a memcache.

So, how about this alternative:

    cache := NetworkCache{Deadline: 5*time.Second}
    Manager{Cache: cache}

Or, when using multiple layers of caches, I would imagine something like this:

    remote := RemoteCache{Deadline: 5*time.Second}
    cache := MemCache{
        Deadline: 50*time.Millisecond,
        Remote: remote,
    }

I'm not convinced, though. Looking for suggestions.

Patch-set: 11
Label: Code-Review=0