acme: refactor internal postJWS for easier signing
Some ACME requests may require to be signed with a key
different from the account key. For instance, a certificate can be
revoked upon a request signed with the key pair in the certificate.

This change will make it easier to implement methods to allow
request signing with different keys.

Change-Id: I635a4d7ad39a63ed9e6823b1af12fbb201c19647